Nearly working Would work if gemini-to-html did not include a require. See RangerMauve/gemini-to-html#2
Ideally we would not load render and parse until we decide we want it. Not sure how we might do this.
Was thinking of just including the scripts before rewriting document contents. Not sure how we can/if we can do this for extensions though.

I suppose agregore-markdown-render does the same.